web-animations/animation-sequence/animation-sequence-synchronization.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationSequence child nodes animation synchronization test</title>
<meta name="assert" content="An animation sequence is a type of animation group that schedules its child animation nodes such that they play in turn following their order in the group">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#animation-sequences">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
async_test(function(t) {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var animationSequence = new AnimationSequence([node1, node2]);
    var player = document.timeline.play(animationSequence);

    player.ready.then(t.step_func(function() {
        animationSequence.player.currentTime = 0;
        assert_equals(window.getComputedStyle(node1.target).top, ANIMATION_TOP_0,
            'Animation should be at the start point for the first node in the sequence');
        assert_equals(window.getComputedStyle(node2.target).top, ANIMATION_TOP_DEFAULT,
            'Animation should not take place for the second node in the sequence');

        animationSequence.player.currentTime = ANIMATION_END_TIME/2;
        assert_equals(window.getComputedStyle(node1.target).top, ANIMATION_TOP_0_5,
            'Animation should take place for the first node in the sequence');
        assert_equals(window.getComputedStyle(node2.target).top, ANIMATION_TOP_DEFAULT,
            'Animation should not take place for the second node in the sequence');

        animationSequence.player.currentTime = ANIMATION_END_TIME;
        assert_equals(window.getComputedStyle(node1.target).top, ANIMATION_TOP_1,
            'Animation should be at the end point for the first node in the sequence');
        assert_equals(window.getComputedStyle(node2.target).top, ANIMATION_TOP_0,
            'Animation should be at the start point for the second node in the sequence');

        animationSequence.player.currentTime = ANIMATION_END_TIME * 1.5;
        assert_equals(window.getComputedStyle(node1.target).top, ANIMATION_TOP_DEFAULT,
            'Animation should be finished for the first node in the sequence');
        assert_equals(window.getComputedStyle(node2.target).top, ANIMATION_TOP_0_5,
            'Animation should take place for the second node in the sequence');

        animationSequence.player.currentTime = ANIMATION_END_TIME * 2;
        assert_equals(window.getComputedStyle(node1.target).top, ANIMATION_TOP_DEFAULT,
            'Animation should be finished for the first node in the sequence');
        assert_equals(window.getComputedStyle(node2.target).top, ANIMATION_TOP_1,
            'Animation should be at the end point for the second node in the sequence');

        t.done();
    }));
}, 'Test that animation takes place in series in AnimationSequence');

async_test(function(t) {
    var node1 = new AnimationGroup([]);
    var node2 = newAnimation(createDiv(this));
    var animationSequence = new AnimationSequence([node1, node2]);
    var player = document.timeline.play(animationSequence);

    player.ready.then(t.step_func(function() {
        animationSequence.player.currentTime = 0;
        assert_equals(window.getComputedStyle(node2.target).top, ANIMATION_TOP_0,
            'Animation should be at the start point for the second node in the sequence');

        animationSequence.player.currentTime = ANIMATION_END_TIME * 0.5;
        assert_equals(window.getComputedStyle(node2.target).top, ANIMATION_TOP_0_5,
            'Animation should take place for the second node in the sequence');

        animationSequence.player.currentTime = ANIMATION_END_TIME;
        assert_equals(window.getComputedStyle(node2.target).top, ANIMATION_TOP_1,
            'Animation should be at the end point for the second node in the sequence');

        t.done();
    }));
}, 'Test that animation takes place in series in AnimationSequence. Test the case when ' +
    'the first node in the sequence is empty group');

async_test(function(t) {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var node3 = new AnimationGroup([node1, node2]);
    var node4 = newAnimation(createDiv(this));
    var animationSequence = new AnimationSequence([node3, node4]);
    var player = document.timeline.play(animationSequence);

    player.ready.then(t.step_func(function() {
        animationSequence.player.currentTime = 0;
        assert_equals(window.getComputedStyle(node1.target).top, ANIMATION_TOP_0,
            'Animation should be at the start point for the first node in ' +
            'the first group in the sequence');
        assert_equals(window.getComputedStyle(node2.target).top, ANIMATION_TOP_0,
            'Animation should be at the start point for the second node in ' +
            'the first group in the sequence');
        assert_equals(window.getComputedStyle(node4.target).top, ANIMATION_TOP_DEFAULT,
            'Animation should not take place for the second node in the sequence');

        animationSequence.player.currentTime = ANIMATION_END_TIME/2;
        assert_equals(window.getComputedStyle(node1.target).top, ANIMATION_TOP_0_5,
            'Animation should take place for the first node in ' +
            'the first group in the sequence');
        assert_equals(window.getComputedStyle(node2.target).top, ANIMATION_TOP_0_5,
            'Animation should take place for the second node in ' +
            'the first group in the sequence');
        assert_equals(window.getComputedStyle(node4.target).top, ANIMATION_TOP_DEFAULT,
            'Animation should not take place for the second node in the sequence');

        animationSequence.player.currentTime = ANIMATION_END_TIME;
        assert_equals(window.getComputedStyle(node1.target).top, ANIMATION_TOP_1,
            'Animation should be at the end point for the first node in ' +
            'the first group in the sequence');
        assert_equals(window.getComputedStyle(node2.target).top, ANIMATION_TOP_1,
            'Animation should be at the end point for the second node in ' +
            'the first group in the sequence');
        assert_equals(window.getComputedStyle(node4.target).top, ANIMATION_TOP_0,
            'Animation should be at the start point for the second node in the sequence');

        animationSequence.player.currentTime = ANIMATION_END_TIME * 1.5;
        assert_equals(window.getComputedStyle(node1.target).top, ANIMATION_TOP_DEFAULT,
            'Animation should be finished for the first node in the first group ' +
            'in the sequence');
        assert_equals(window.getComputedStyle(node2.target).top, ANIMATION_TOP_DEFAULT,
                'Animation should be finished for the second node in the first group ' +
                'in the sequence');
        assert_equals(window.getComputedStyle(node4.target).top, ANIMATION_TOP_0_5,
            'Animation should take place for the second node in the sequence');

        animationSequence.player.currentTime = ANIMATION_END_TIME * 2;
        assert_equals(window.getComputedStyle(node1.target).top, ANIMATION_TOP_DEFAULT,
            'Animation should be finished for the first node in the first group ' +
            'in the sequence');
        assert_equals(window.getComputedStyle(node2.target).top, ANIMATION_TOP_DEFAULT,
            'Animation should be finished for the second node in the first group ' +
            'in the sequence');
        assert_equals(window.getComputedStyle(node2.target).top, ANIMATION_TOP_1,
            'Animation should be at the end point for the second node in the sequence');

        t.done();
    }));
}, 'Test that animation takes place in series in AnimationSequence. ' +
    'Test the case when the first node in the sequence is AnimationGroup');

async_test(function(t) {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var node3 = newAnimation(createDiv(this));
    var node4 = new AnimationGroup([node1, node2]);
    var animationSequence = new AnimationSequence([node3, node4]);
    var player = document.timeline.play(animationSequence);

    player.ready.then(t.step_func(function() {
        animationSequence.player.currentTime = 0;
        assert_equals(window.getComputedStyle(node3.target).top, ANIMATION_TOP_0,
            'Animation should be at the start point for the first node in the sequence');
        assert_equals(window.getComputedStyle(node1.target).top, ANIMATION_TOP_DEFAULT,
            'Animation should not take place for nodes in the second node in the sequence');
        assert_equals(window.getComputedStyle(node2.target).top, ANIMATION_TOP_DEFAULT,
            'Animation should not take place for nodes in the second node in the sequence');

        animationSequence.player.currentTime = ANIMATION_END_TIME/2;
        assert_equals(window.getComputedStyle(node3.target).top, ANIMATION_TOP_0_5,
            'Animation should take place for the first node in the sequence');
        assert_equals(window.getComputedStyle(node1.target).top, ANIMATION_TOP_DEFAULT,
            'Animation should not take place for nodes in the second node in the sequence');
        assert_equals(window.getComputedStyle(node2.target).top, ANIMATION_TOP_DEFAULT,
            'Animation should not take place for nodes in the second node in the sequence');

        animationSequence.player.currentTime = ANIMATION_END_TIME;
        assert_equals(window.getComputedStyle(node3.target).top, ANIMATION_TOP_1,
            'Animation should be at the end point for the first node in the sequence');
        assert_equals(window.getComputedStyle(node1.target).top, ANIMATION_TOP_0,
            'Animation should be at the start point for nodes in the second node ' +
            'in the sequence');
        assert_equals(window.getComputedStyle(node2.target).top, ANIMATION_TOP_0,
            'Animation should be at the start point for nodes in the second node ' +
            'in the sequence');

        animationSequence.player.currentTime = ANIMATION_END_TIME * 1.5;
        assert_equals(window.getComputedStyle(node3.target).top, ANIMATION_TOP_DEFAULT,
            'Animation should be finished for the first node in the sequence');
        assert_equals(window.getComputedStyle(node1.target).top, ANIMATION_TOP_0_5,
            'Animation should take place for the nodes in the second node in the sequence');
        assert_equals(window.getComputedStyle(node2.target).top, ANIMATION_TOP_0_5,
            'Animation should take place for the nodes in the second node in the sequence');

        animationSequence.player.currentTime = ANIMATION_END_TIME * 2;
        assert_equals(window.getComputedStyle(node3.target).top, ANIMATION_TOP_DEFAULT,
            'Animation should be finished for the first node in the sequence');
        assert_equals(window.getComputedStyle(node1.target).top, ANIMATION_TOP_1,
            'Animation should be at the end point for the second node in the sequence');
        assert_equals(window.getComputedStyle(node2.target).top, ANIMATION_TOP_1,
            'Animation should be at the end point for the second node in the sequence');

        t.done();
    }));
}, 'Test that animation takes place in series in AnimationSequence. ' +
    'Test the case when the second node in the sequence is AnimationGroup');
</script>
</body>
